Refine defaults when linking
Originally I set the default min year to 2010 under the idea that
people are more interested in recent games. But the problem with that
is when we link back to the search with a specific term it might
filter out what we are looking for. For example linking to
`?query=Uncharted` only shows the 3rd title even though I expect all
three.

With this update we refine the default so that if there was no initial
query I still use the 2010 default. But if there was a initial query
then it uses the min year.