Bump deps
This doesn't really get rid of the flagged packages in `yarn audit`
because parcel depends on vuln packages. But since this is all part of
the build system there isn't any real vuln so just marked the alerts off
in Github and pushing it up with everything bumped up the most it can.

Looks like Parcel v2 RC is out a few days ago so once that rolls to
production we can get rid of those items more proper.